+++ src/config/docs.md
@@ -10,14 +10,19 @@ This page is automatically generated by `./risedev generate-example-config`
 | enable_barrier_read |  | false |
 | enable_spill | Enable the spill out to disk feature for batch queries. | true |
 | frontend_compute_runtime_worker_threads | frontend compute runtime worker threads | 4 |
-| frontend_max_running_message_bytes | Total maximum memory allowed for running message in the frontend. | 1073741824 |
-| frontend_throttling_filter_max_bytes | A message exceeding this threshold will always be throttled for memory constraints. | 1073741824 |
-| frontend_throttling_filter_min_bytes | A message under this threshold will not be throttled for memory constraints. | 1048576 |
 | mask_worker_temporary_secs | This is the secs used to mask a worker unavailable temporarily. | 30 |
 | max_batch_queries_per_frontend_node | This is the max number of batch queries per frontend node. |  |
 | redact_sql_option_keywords | Keywords on which SQL option redaction is based in the query log. A SQL option with a name containing any of these keywords will be redacted. | ["credential", "key", "password", "private", "secret", "token"] |
 | statement_timeout_in_sec | Timeout for a batch query in seconds. | 3600 |
 | worker_threads_num | The thread number of the batch task runtime in the compute node. The default value is decided by `tokio`. |  |
+
+## frontend
+
+| Config | Description | Default |
+|--------|-------------|---------|
+| max_single_query_size_bytes | A query of size exceeding this threshold will always be rejected due to memory constraints. | 1073741824 |
+| max_total_query_size_bytes | Total memory constraints for running queries. | 1073741824 |
+| min_single_query_size_bytes | A query of size under this threshold will never be rejected due to memory constraints. | 1048576 |
 
 ## meta
